feat: planar バックエンド撤去・globe 単一化 (#414)
Phase 6: terrainEngine を globe 単一バックエンドへ統合する。

- TerrainEngine 型から "planar" を撤去し "globe" 単一へ
- planar 本体 (scenes/default.ts) と createScene.ts を削除
- 共有 export を新規 scenes/sceneContract.ts へ退避
  (METERS_PER_DEGREE_LAT / MarkerContext / DefaultSceneController /
   DefaultSceneInitOptions)
- urlState/jpmapTerrain/各デモの planar 分岐・artillery フォールバックを撤去
- ?terrainEngine=planar は無効値→globe 既定にフォールバック (互換不要)
- JAPAN_BOUNDS は globe が実利用のため存置
- 単体テストの planar ケースを globe へ更新
- VR (validation.spec.ts) を globe 単一化。Compass reset を GeospatialCamera の
  yaw/pitch 判定へ修正し reset 後の waitForTerrainStable で安定化
- 削除済み scenes/default.ts への dangling コメント参照を整理

lint / typecheck / test:unit(1367) / VR(再ベースライン) 全緑。
