--- src/createScene.ts
@@ -1,22 +1,0 @@
-import type { Scene } from "@babylonjs/core/scene";
-
-// Change this import to check other scenes
-import { DefaultScene, type DefaultSceneInitOptions } from "./scenes/default";
-import { AbstractEngine } from "@babylonjs/core/Engines/abstractEngine";
-
-export interface CreateSceneClass {
-    createScene: (
-        engine: AbstractEngine,
-        canvas: HTMLCanvasElement,
-        options?: DefaultSceneInitOptions,
-    ) => Promise<Scene>;
-    preTasks?: Promise<unknown>[];
-}
-
-export interface CreateSceneModule {
-    default: CreateSceneClass;
-}
-
-export const getSceneModule = (): CreateSceneClass => {
-    return new DefaultScene();
-}
