+++ tests/index.unit.spec.ts
@@ -27,8 +27,8 @@ describe("resolveTerrainEngine", () => {
         expect(resolveTerrainEngine("?terrainEngine=globe")).toBe("globe");
     });
 
-    it("?terrainEngine=planar → 'planar'", () => {
-        expect(resolveTerrainEngine("?terrainEngine=planar")).toBe("planar");
+    it("?terrainEngine=planar（撤去済み）→ undefined", () => {
+        expect(resolveTerrainEngine("?terrainEngine=planar")).toBeUndefined();
     });
 
     it("未指定 → undefined（lib 既定 globe にフォールバック, #413）", () => {
